Make some plugins optional, move ps to plugin
