+++ Cargo.toml
@@ -50,7 +50,6 @@ git2 = { version = "0.10.0", default_features = false }
 dirs = "2.0.2"
 glob = "0.3.0"
 ctrlc = "3.1.3"
-ptree = "0.2"
 surf = "1.0.2"
 url = "2.1.0"
 roxmltree = "0.7.0"
@@ -62,29 +61,36 @@ subprocess = "0.1.18"
 mime = "0.3.13"
 regex = "1.2.1"
 pretty-hex = "0.1.0"
-neso = { version = "0.5.0", optional = true }
 hex = "0.3.2"
-crossterm = "0.10.2"
 tempfile = "3.1.0"
-image = { version = "0.22.1", default_features = false, features = ["png_codec", "jpeg"] }
 semver = "0.9.0"
-uuid = {version = "0.7.4", features = [ "v4", "serde" ]}
-syntect = "3.2.0"
-onig_sys = "=69.1.0"
-heim = "0.0.7"
 which = "2.0.1"
-battery = "0.7.4"
+uuid = {version = "0.7.4", features = [ "v4", "serde" ]}
 textwrap = {version = "0.11.0", features = ["term_size"]}
-rawkey = {version = "0.1.2", optional = true }
-clipboard = {version = "0.5", optional = true }
 shellexpand = "1.0.0"
 futures-timer = "0.3.0"
 pin-utils = "0.1.0-alpha.4"
 num-bigint = { version = "0.2.2", features = ["serde"] }
 bigdecimal = { version = "0.1.0", features = ["serde"] }
 
+neso = { version = "0.5.0", optional = true }
+crossterm = { version = "0.10.2", optional = true }
+syntect = {version = "3.2.0", optional = true }
+onig_sys = {version = "=69.1.0", optional = true }
+heim = {version = "0.0.7", optional = true }
+battery = {version = "0.7.4", optional = true }
+rawkey = {version = "0.1.2", optional = true }
+clipboard = {version = "0.5", optional = true }
+ptree = {version = "0.2", optional = true }
+image = { version = "0.22.1", default_features = false, features = ["png_codec", "jpeg"], optional = true }
+
 [features]
 raw-key = ["rawkey", "neso"]
+textview = ["syntect", "onig_sys", "crossterm"]
+binaryview = ["image", "crossterm"]
+sys = ["heim", "battery"]
+ps = ["heim"]
+all = ["raw-key", "textview", "binaryview", "sys", "ps", "clipboard", "ptree"]
 
 [dependencies.rusqlite]
 version = "0.20.0"
@@ -128,18 +134,27 @@ path = "src/plugins/skip.rs"
 [[bin]]
 name = "nu_plugin_sys"
 path = "src/plugins/sys.rs"
+required-features = ["sys"]
+
+[[bin]]
+name = "nu_plugin_ps"
+path = "src/plugins/ps.rs"
+required-features = ["ps"]
 
 [[bin]]
 name = "nu_plugin_tree"
 path = "src/plugins/tree.rs"
+required-features = ["tree"]
 
 [[bin]]
 name = "nu_plugin_binaryview"
 path = "src/plugins/binaryview.rs"
+required-features = ["binaryview"]
 
 [[bin]]
 name = "nu_plugin_textview"
 path = "src/plugins/textview.rs"
+required-features = ["textview"]
 
 [[bin]]
 name = "nu"
