+++ src/commands.rs
@@ -39,7 +39,6 @@ pub(crate) mod pick;
 pub(crate) mod plugin;
 pub(crate) mod post;
 pub(crate) mod prev;
-pub(crate) mod ps;
 pub(crate) mod pwd;
 pub(crate) mod reject;
 pub(crate) mod reverse;
@@ -106,7 +105,6 @@ pub(crate) use open::Open;
 pub(crate) use pick::Pick;
 pub(crate) use post::Post;
 pub(crate) use prev::Previous;
-pub(crate) use ps::PS;
 pub(crate) use pwd::PWD;
 pub(crate) use reject::Reject;
 pub(crate) use reverse::Reverse;
